Move ModelTest to package no.nva.unit -> no.unit.nva.model
