+++ src/test/java/no/unit/nva/model/ContributorTest.java
@@ -3,20 +3,10 @@ package no.unit.nva.model;
 import no.unit.nva.model.exceptions.MalformedContributorException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.function.Executable;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.CsvSource;
 
-import java.net.URI;
 import java.util.Collections;
 
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
-import no.unit.nva.ModelTest;
-import no.unit.nva.model.exceptions.MalformedContributorException;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-
-import java.util.Collections;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
