+++ src/test/java/no/unit/nva/model/contexttypes/DegreeTest.java
@@ -3,7 +3,7 @@ package no.unit.nva.model.contexttypes;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import no.unit.nva.ModelTest;
+import no.unit.nva.model.ModelTest;
 import no.unit.nva.model.exceptions.InvalidIsbnException;
 import nva.commons.utils.JsonUtils;
 import org.junit.jupiter.api.DisplayName;
